+++ Packages/com.fang.framework/Samples~/MinimalSample/HeroWorldObject.cs
@@ -1,6 +1,0 @@
-namespace Fang.Framework.MinimalSample
-{
-    public sealed class HeroWorldObject : WorldObject<HeroController, HeroData, HeroConfigDataSo>
-    {
-    }
-}